pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symphony-orchestra"
version = "1.0.0-beta.3"
description = "Lightweight container and job orchestrator"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "aiosqlite==0.22.1",
    "fastapi==0.128.0",
    "grpcio==1.76.0",
    "orjson==3.11.5",
    "pydantic==2.12.5",
    "PyYAML==6.0.3",
    "uvicorn==0.40.0",
    "cryptography==44.0.0",
    "nvidia-ml-py3==7.352.0",
    "autoflake==2.3.1",
    "black==25.12.0",
    "grpcio-tools==1.76.0",
    "isort==7.0.0",
    "pyflakes==3.4.0",
    "click==8.3.1",
    "packaging==25.0",
    "pathspec==0.12.1",
    "platformdirs==4.5.1",
    "mypy_extensions==1.1.0",
    "pytokens==0.3.0",
    "loguru==0.7.3"
]
[project.scripts]
symphony = "symphony.__main__:main"
[project.readme]
file = "README.md"
content-type = "text/markdown"
[project.urls]
Homepage = "https://github.com/ttheew/symphony"
Source = "https://github.com/ttheew/symphony"
Issues = "https://github.com/ttheew/symphony/issues"
Documentation = "https://github.com/ttheew/symphony#readme"
[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["symphony*"]
exclude = [
    "symphony-conductor-webui",
    "symphony-conductor-webui.*"
]
[tool.setuptools.package-data]
symphony = [
    "conductor/ui/dist/**",
    "docs/**"
]